requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Pillow==8.3.1
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.0.1
gunicorn==20.1.0
dnspython==2.1.0
bcrypt==4.0.1
email-validator==2.0.0
Jinja2==3.1.2 